requirements.txt
streamlit
pandas
pyzipper
openpyxl
langchain-community
openai
python-dotenv
tabulate
plotly